requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.2.0
boto3==1.34.11
botocore==1.34.11
celery==5.3.6
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
Django==5.0
django-celery-beat==2.6.0
django-redis==5.4.0
django-role-permissions==3.2.0
django-storages==1.14.2
django-timezone-field==6.1.0
gunicorn==21.2.0
idna==3.6
jmespath==1.0.1
kombu==5.3.5
numpy==1.26.2
packaging==23.2
pandas==2.1.4
Pillow==10.1.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pypng==0.20220715.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
qrcode==7.4.2
redis==5.0.1
requests==2.31.0
s3transfer==0.10.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.0.2
vine==5.1.0
wcwidth==0.2.13
WooCommerce==3.0.0
